GUI: Add ability to migrate resources (bnc#593005)
This changeset also causes the wait spinners to only be active for the
duration of the AJAX request that performs a given op.  The alternative
is to risk having spinners running indefinitely, which is slightly worse
(either case confuses the user about current status, but for now that's
where we are).