GUI: explorer: Show diff link on every line
Previously we were showing diff links for pairs of pe inputs, which
stupidly meant you could diff item 1 with item 2, and 3 with 4, but
not 2 with 3.  This change just puts a diff link on every line except
the last, with the diff link diffing the current line with the next
line.  This is marginally visually less clear, but acutally lets you
diff the full set of pe inputs.

Signed-off-by: Tim Serong <tserong@suse.com>