Build: Touch POT file back to ancient date (ensures correct generation)
It's possible to end up with the POT file having a newer mtime than the
various source files we need to search for translations (say, if you've
been checking out different versions of it).  If the POT file mtime is
too new, text strings from "older" source won't be pulled in, hence the
`touch` invocation.

Signed-off-by: Tim Serong <tserong@suse.com>