In-progress wizard work
Calling the cluster script seems to work, but there a lot of things
missing: Right now, there is no deeper interaction between the wizard
and the cluster script.

TODO:

1. appropriate parameters to the cluster script (matching the parameters
   to the wizard)
2. hawk updating the statefile with configured values before Validate
3. calling the cluster script as root (and requesting root privileges)
4. allowing multiple exportfs elements in the wizard.

My initial idea for 4) was to add something like

In the exportfs template, an attribute that says that this step can be
done multiple times (no good idea for an attribute name here)

In the <crm_script> section, use

<for template="exportfs">
...
</for>
which should expand to all the instances of the exportfs template in
sequence.