GUI: status: Add 'elsewhere' state for tickets not granted to current site (bnc#877482)
The summary and tree view will now show tickets known to be granted to
some other site as "Elsewhere".  This is a variant on Revoked, but makes
it obvious that booth thinks the ticket is somewhere else.  If booth
isn't available, the ticket will just show as Revoked.

The dashboard still only has granted and revoked, because you can view
multiple clusters at once, so the Elsewhere state doesn't make sense in
that view.

Signed-off-by: Tim Serong <tserong@suse.com>