Scripts: Send SIGINT to stop puma
systemd requires the ExecStop= process to wait for the process
to stop. pumactl doesn't do this correctly, so the previous
method for stopping it would fail during restart since the
stop didn't wait for puma to actually terminate before restarting
it.

Puma listens to SIGINT and performs a graceful shutdown, so
sending SIGINT to stop it works better.