Monitor: Implement /monitor as live controller (bsc#947748)
The frontend uses /monitor to implement a kind of server-
sent events. To enable this with plain Puma + Rails 4, we
use a Live controller.

This unfortunately has some radical implications:

1. The app must be thread-safe.

2. We have to disable auto-loading in development.

Another option is to turn off the long-running /monitor
call in development. Perhaps once this is stable, that
is a nicer option. Having to restart the app all the time
sucks pretty bad.

* Require session for /monitor

* Trigger updated.hawk.monitor event from simulator