Replace Simulator with Batch Mode (fate#318298) (fate#317834) (fate#318314)
This replaces the previous Simulator interface with Batch Mode:

When enabling Batch Mode, a control panel bar appears at the top
of the interface. From this, it's possible to inject simulator
events. The simulator is always run on any configuration change
or event injection.

The results of simulation can be viewed as before, and there is
also a diff view of the configuration changes.

The configuration changes can then either be applied to the
real configuration, or discarded.