UI: Correct font-awesome icon asset paths (bsc#974435)
The font-awesome paths need to include the asset compilation
hash if they are to actually use the compilated versions. For
some reason, the uncompiled fonts were still being shipped
with our package so things worked regardless, but as I fixed
the asset pipeline, this unfortunately broke font-awesome.

This commit makes font-awesome use the same path rewriting as
used by the bootstrap glyphicons.