Dashboard: Cycle through node list if credentials are set
If we have a username and password already, we'll use that. But if
this was the local node, we don't have credentials set so we just
try to reconnect locally.

TODO: Add form for entering username/password to the local cluster
so that we can connect to other nodes in the local cluster as well.