Set HttpOnly by HAWK_COOKIE_HTTP_ONLY=true
The flag HttpOnly was unconditional. It was set fot all cookies.
This commit gives you a choice. To set the HttpOnly flag you
need to pass the environment variable HAWK_COOKIE_HTTP_ONLY=true.
The simplest way to do it is to add it to the /etc/sysconfig/hawk file,
for example
$ echo "HAWK_COOKIE_HTTP_ONLY=true" >> /etc/sysconfig/hawk
$ systemctl restart hawk-backend