Dev: rename 'Score' to 'Kind'
This is the next iteration of renaming 'score' -> 'kind'.
Here we replace all occurences of the 'score' visible to
the user. So, an English version hawk will display all
correctly ('kind'). However
1) The part hidden from the user: backend, html and js tags,
etc. are still 'score'. Renaming those variables is an epic,
so maybe it's better not to touch them.
2) Renaming was mostly done automatically,
'Score of INFINITY' becomes 'Kind of INFINITY'
3) In the other languages the 'score' remains 'score'.